Refs #24682: BaseImage - remove unused injection, move interfaces into appropriate location
