+++ src/modules/budsies/types/image-aspect-ratio-spec.interface.ts
